+++ compat/jansson/utf.c
@@ -167,7 +167,7 @@ int utf8_check_string(const char *string, int length)
     int i;
 
     if(length == -1)
-        length = strlen(string);
+        length = (int)strlen(string);
 
     for(i = 0; i < length; i++)
     {
